requirements.txt
annotated-types==0.6.0
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
crispy-bootstrap5==0.7
cryptography==39.0.1
defusedxml==0.7.1
dj-database-url==2.1.0
dj-email-url==1.0.6
Django==4.2.5
django-allauth==0.57.0
django-cache-url==3.4.4
django-crispy-forms==2.0
django-debug-toolbar==4.2.0
django-htmx==1.17.0
django-qr-code==3.1.1
environs==9.5.0
gunicorn==21.2.0
idna==3.4
marshmallow==3.20.1
oauthlib==3.2.2
packaging==23.1
Pillow==10.1.0
psycopg==3.1.12
psycopg-binary==3.1.12
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.6.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
qrcode-artistic==3.0.0
requests==2.28.2
requests-oauthlib==1.3.1
segno==1.5.2
sentry-sdk==1.37.1
sqlparse==0.4.3
stripe==4.2.0
typing_extensions==4.8.0
urllib3==1.26.14
whitenoise==6.5.0
